+++ library/CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.0.5
+
+- Add progress reporting support
+- Remove leaked console.log
+
 ## 1.0.4
 
 - Fix incorrect shapes filtering on creating boolean shapes within components